refactor: break down gallery and viewer into sub-components
