+++ src/components/organisms/gallery/Gallery.tsx
@@ -1,12 +1,13 @@
-import { Button, Image } from '@components/atoms';
-import { ButtonView, IconName } from '@enums';
 import type { GalleryProps } from '@types';
 import { getClassName } from '@utils';
-import { type CSSProperties, type FC, useCallback, useRef, useState } from 'react';
+import { type CSSProperties, type FC, useRef } from 'react';
 
-import FullscreenViewer from './FullscreenViewer';
+import CarouselNavigation from './carousel-navigation/CarouselNavigation';
+import FullscreenViewer from './full-screen-viewer/FullscreenViewer';
 import './Gallery.css';
-import { useCarouselScroll } from './use-carousel-scroll';
+import GalleryImageItem from './gallery-image-item/GalleryImageItem';
+import { useCarouselScroll } from './hooks/use-carousel-scroll';
+import { useFullscreenViewer } from './hooks/use-fullscreen-viewer';
 
 const Gallery: FC<GalleryProps> = ({
   images = [],
@@ -22,38 +23,8 @@ const Gallery: FC<GalleryProps> = ({
 }) => {
   const containerRef = useRef<HTMLDivElement>(null);
 
-  // Fullscreen viewer state management
-  const [isViewerOpen, setIsViewerOpen] = useState(false);
-  const [currentImageIndex, setCurrentImageIndex] = useState(0);
-  const triggerRef = useRef<HTMLElement | null>(null);
-
-  const openViewer = useCallback((index: number, element: HTMLElement) => {
-    triggerRef.current = element;
-    setCurrentImageIndex(index);
-    setIsViewerOpen(true);
-  }, []);
-
-  const closeViewer = useCallback(() => {
-    setIsViewerOpen(false);
-    // Focus restoration happens after animation completes
-    setTimeout(() => {
-      triggerRef.current?.focus();
-      triggerRef.current = null;
-    }, 300);
-  }, []);
-
-  // Navigation handler for FullscreenViewer (used in task 6.4)
-  const handleNavigate = useCallback((index: number) => {
-    setCurrentImageIndex(index);
-  }, []);
-
-  // Keyboard handler for image activation (Enter/Space)
-  const handleImageKeyDown = (e: React.KeyboardEvent<HTMLDivElement>, index: number) => {
-    if (e.key === 'Enter' || e.key === ' ') {
-      e.preventDefault();
-      openViewer(index, e.currentTarget);
-    }
-  };
+  const { isViewerOpen, currentImageIndex, openViewer, closeViewer, handleNavigate } =
+    useFullscreenViewer();
 
   const { showPrev, showNext, scrollForward, scrollBackward } = useCarouselScroll(
     containerRef,
@@ -72,10 +43,6 @@ const Gallery: FC<GalleryProps> = ({
     '--gallery-rows': rows,
     '--gallery-gap': gap,
   } as CSSProperties;
-  const itemClass = getClassName([
-    'cl-gallery__item',
-    { 'cl-gallery__item_fullscreen': enableFullscreen },
-  ]);
 
   return (
     <section
@@ -91,40 +58,22 @@ const Gallery: FC<GalleryProps> = ({
         {...(layout === 'carousel' ? { 'aria-roledescription': 'carousel' } : {})}
       >
         {images.map((item, index) => (
-          <div
+          <GalleryImageItem
             key={item.src}
-            className={itemClass}
-            tabIndex={enableFullscreen ? 0 : undefined}
-            role={enableFullscreen ? 'button' : undefined}
-            aria-haspopup={enableFullscreen ? 'dialog' : undefined}
-            onClick={enableFullscreen ? (e) => openViewer(index, e.currentTarget) : undefined}
-            onKeyDown={enableFullscreen ? (e) => handleImageKeyDown(e, index) : undefined}
-          >
-            <Image {...item} />
-          </div>
+            item={item}
+            index={index}
+            enableFullscreen={enableFullscreen}
+            onOpenViewer={openViewer}
+          />
         ))}
       </div>
       {layout === 'carousel' && (
-        <>
-          <Button
-            view={ButtonView.Icon}
-            icon={IconName.ChevronLeft}
-            label="Previous images"
-            className="cl-gallery__button cl-gallery__button_prev"
-            onClick={scrollBackward}
-            style={{ visibility: showPrev ? 'visible' : 'hidden' }}
-            aria-hidden={!showPrev}
-          />
-          <Button
-            view={ButtonView.Icon}
-            icon={IconName.ChevronRight}
-            label="Next images"
-            className="cl-gallery__button cl-gallery__button_next"
-            onClick={scrollForward}
-            style={{ visibility: showNext ? 'visible' : 'hidden' }}
-            aria-hidden={!showNext}
-          />
-        </>
+        <CarouselNavigation
+          showPrev={showPrev}
+          showNext={showNext}
+          onScrollBackward={scrollBackward}
+          onScrollForward={scrollForward}
+        />
       )}
       {enableFullscreen && (
         <FullscreenViewer
