+++ src/components/organisms/gallery/gallery-image-item/GalleryImageItem.css
@@ -1,0 +1,7 @@
+.cl-gallery__item {
+  overflow: hidden;
+}
+
+.cl-gallery__item_fullscreen {
+  cursor: pointer;
+}
