+++ src/components/organisms/gallery/hooks/use-carousel-scroll.ts
