+++ src/components/organisms/gallery/viewer-overlay/ViewerOverlay.css
@@ -1,0 +1,10 @@
+.cl-fullscreen-viewer__overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
